--- urbackupserver/www2/src/features/settings/SettingsServer/SettingsServer.module.css
@@ -1,35 +1,0 @@
-.container {
-  --border-color: var(--colorNeutralStroke2);
-  --flow-space: 2em;
-
-  width: 74ch;
-  margin: 0 auto;
-  padding-block-end: --spacing;
-}
-
-.card {
-  display: "flex";
-  flex-direction: "column";
-  background: var(--colorNeutralCardBackground);
-  padding: var(--spacingL);
-  border-radius: var(--borderRadiusLarge);
-  border-width: 1px;
-  border-style: solid;
-  border-color: var(--border-color);
-}
-
-.form > :has(label) {
-  padding-block: 0.5em;
-}
-
-.form > :first-child:has(label) {
-  padding-block-start: 0em;
-}
-
-.form > :last-child:has(label) {
-  padding-block-end: 0em;
-}
-
-.form > :has(label):not(:last-of-type) {
-  border-block-end: 1px solid var(--border-color);
-}
